moved all cli runners to their relevant test modules. Closes #52
