+++ tests/providers/test_join.py
@@ -1,4 +1,7 @@
+import os
+
 import pytest
+from click.testing import CliRunner
 
 from notifiers import get_notifier
 from notifiers.exceptions import BadArguments
@@ -36,3 +39,27 @@ class TestJoin:
         data = {'message': 'foo'}
         rsp = p.notify(**data)
         rsp.raise_on_errors()
+
+
+class TestJoinCLI:
+    """Test join specific CLI"""
+
+    def test_join_devices_negative(self):
+        from notifiers_cli.providers.join import devices
+        runner = CliRunner()
+        result = runner.invoke(devices, ['bad_token'])
+        assert result.exit_code == -1
+        assert not result.output
+
+    @pytest.mark.skip('tests fail due to no device connected')
+    @pytest.mark.online
+    def test_join_updates_positive(self):
+        from notifiers_cli.providers.join import devices
+        token = os.environ.get('NOTIFIERS_JOIN_APIKEY')
+        assert token
+
+        runner = CliRunner()
+        result = runner.invoke(devices, [token])
+        assert result.exit_code == 0
+        replies = ['You have no devices associated with this apikey', 'Device name: ']
+        assert any(reply in result.output for reply in replies)
