+++ tests/providers/test_pushbullet.py
@@ -1,4 +1,7 @@
+import os
+
 import pytest
+
 
 from notifiers import get_notifier
 from notifiers.exceptions import BadArguments
@@ -45,3 +48,26 @@ class TestPushbullet:
         }
         rsp = p.notify(**data)
         rsp.raise_on_errors()
+
+
+class TestPushbulletCLI:
+    """Test pushbullet specific CLI"""
+
+    def test_pushbullet_devices_negative(self):
+        from notifiers_cli.providers.pushbullet import devices
+        runner = CliRunner()
+        result = runner.invoke(devices, ['bad_token'])
+        assert result.exit_code == -1
+        assert not result.output
+
+    @pytest.mark.online
+    def test_telegram_updates_positive(self):
+        from notifiers_cli.providers.pushbullet import devices
+        token = os.environ.get('NOTIFIERS_PUSHBULLET_TOKEN')
+        assert token
+
+        runner = CliRunner()
+        result = runner.invoke(devices, [token])
+        assert result.exit_code == 0
+        replies = ['You have no devices associated with this token', 'Nickname: ']
+        assert any(reply in result.output for reply in replies)
