+++ tests/test_cli.py
@@ -99,124 +99,3 @@ class TestCLI:
         result = runner.invoke(notify, [], input='foo')
         assert result.exit_code == 0
         assert not result.output
-
-
-class TestGitterCLI:
-    """Test gitter specific CLI commands"""
-
-    def test_gitter_rooms_negative(self):
-        from notifiers_cli.providers.gitter import rooms
-        runner = CliRunner()
-        result = runner.invoke(rooms, ['bad_token'])
-        assert result.exit_code == -1
-        assert not result.output
-
-    @pytest.mark.online
-    def test_gitter_rooms_positive(self):
-        from notifiers_cli.providers.gitter import rooms
-        token = os.environ.get('NOTIFIERS_GITTER_TOKEN')
-        assert token
-
-        runner = CliRunner()
-        result = runner.invoke(rooms, [token])
-        assert result.exit_code == 0
-        assert 'notifiers/testing' in result.output
-
-    @pytest.mark.online
-    def test_gitter_rooms_with_query(self):
-        from notifiers_cli.providers.gitter import rooms
-        token = os.environ.get('NOTIFIERS_GITTER_TOKEN')
-        assert token
-
-        runner = CliRunner()
-        result = runner.invoke(rooms, [token, '-q', 'notifiers/testing'])
-        assert result.exit_code == 0
-        assert 'notifiers/testing' in result.output
-
-
-class TestTelegramCLI:
-    """Test telegram specific CLI"""
-
-    def test_telegram_updates_negative(self):
-        from notifiers_cli.providers.telegram import updates
-        runner = CliRunner()
-        result = runner.invoke(updates, ['bad_token'])
-        assert result.exit_code == -1
-        assert not result.output
-
-    @pytest.mark.online
-    def test_telegram_updates_positive(self):
-        from notifiers_cli.providers.telegram import updates
-        token = os.environ.get('NOTIFIERS_TELEGRAM_TOKEN')
-        assert token
-
-        runner = CliRunner()
-        result = runner.invoke(updates, [token])
-        assert result.exit_code == 0
-        replies = ['Bot has not active chats! Send it ANY message and try again', 'Chat ID:']
-        assert any(reply in result.output for reply in replies)
-
-
-class TestPushbulletCLI:
-    """Test pushbullet specific CLI"""
-
-    def test_pushbullet_devices_negative(self):
-        from notifiers_cli.providers.pushbullet import devices
-        runner = CliRunner()
-        result = runner.invoke(devices, ['bad_token'])
-        assert result.exit_code == -1
-        assert not result.output
-
-    @pytest.mark.online
-    def test_telegram_updates_positive(self):
-        from notifiers_cli.providers.pushbullet import devices
-        token = os.environ.get('NOTIFIERS_PUSHBULLET_TOKEN')
-        assert token
-
-        runner = CliRunner()
-        result = runner.invoke(devices, [token])
-        assert result.exit_code == 0
-        replies = ['You have no devices associated with this token', 'Nickname: ']
-        assert any(reply in result.output for reply in replies)
-
-
-class TestJoinCLI:
-    """Test join specific CLI"""
-
-    def test_join_devices_negative(self):
-        from notifiers_cli.providers.join import devices
-        runner = CliRunner()
-        result = runner.invoke(devices, ['bad_token'])
-        assert result.exit_code == -1
-        assert not result.output
-
-    @pytest.mark.skip('tests fail due to no device connected')
-    @pytest.mark.online
-    def test_join_updates_positive(self):
-        from notifiers_cli.providers.join import devices
-        token = os.environ.get('NOTIFIERS_JOIN_APIKEY')
-        assert token
-
-        runner = CliRunner()
-        result = runner.invoke(devices, [token])
-        assert result.exit_code == 0
-        replies = ['You have no devices associated with this apikey', 'Device name: ']
-        assert any(reply in result.output for reply in replies)
-
-
-class TestHipchatCLI:
-    """Test hipchat specific CLI"""
-
-    def test_hipchat_rooms_negative(self):
-        from notifiers_cli.providers.hipchat import rooms
-        runner = CliRunner()
-        result = runner.invoke(rooms, ['bad_token'])
-        assert result.exit_code == -1
-        assert not result.output
-
-    def test_hipchat_users_negative(self):
-        from notifiers_cli.providers.hipchat import users
-        runner = CliRunner()
-        result = runner.invoke(users, ['bad_token'])
-        assert result.exit_code == -1
-        assert not result.output
